move t_time.cpp/h to common folder
